DFWU (DDNS Firewall Update) Gist to Repo
Server-side for DFWU (DDNS Firewall Update) brought from GitHub Gist to
GitHub Repository.

Up next is the client that's on machines (servers, clients, embedded)
which dial in to server(s) running DFWU.  Added bonus will be an
installer that takes care of placing DFWU entries in appropriate
locations and/or an Ansible playbook to match it, so that everything is
more turnkey.